Move anonymous TomcatEmbeddedServletContainerFactory and associated context customiser to separate class
