+++ api/cas-server-core-api-configuration-model/src/main/java/org/apereo/cas/configuration/model/core/web/tomcat/CasEmbeddedApacheTomcatClusteringProperties.java
@@ -3,6 +3,7 @@ package org.apereo.cas.configuration.model.core.web.tomcat;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
 import org.apereo.cas.configuration.support.RequiredProperty;
 import org.apereo.cas.configuration.support.RequiresModule;
 
@@ -138,4 +139,8 @@ public class CasEmbeddedApacheTomcatClusteringProperties implements Serializable
      * Enable tomcat session clustering.
      */
     private boolean enabled;
+
+    public boolean isSessionClusteringEnabled() {
+        return isEnabled() && StringUtils.isNotBlank(getClusterMembers());
+    }
 }
